Создаст файл который генерирует массив
