+++ js/util.js
@@ -10,3 +10,4 @@ const getRandomInteger = (a, b) => {
 const getRandomArrayElement = (elements) => elements[getRandomInteger(0, elements.length - 1)];
 
 export{getRandomArrayElement};
+export{getRandomInteger};
